+++ src/Data/TypedEncoding/Instances/Support/Helpers.hs
@@ -20,7 +20,6 @@ import           Data.TypedEncoding.Combinators.Unsafe
 import           Data.TypedEncoding.Common.Class.IsStringR 
 import           GHC.TypeLits
 
--- TODO v0.3 some of this needs to move to call side exported combinators
 
 -- $setup
 -- >>> :set -XTypeApplications